# specs/001-crud-audit-viewmodel-api/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: rule-condition-engine API
  version: 0.2.0
  description: >
    Uniform record-management contract (persons, cases, rules, person-case links) plus the
    rule-specific operations that are not CRUD. See crud-contract.md for the rules that every
    resource path below inherits identically.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/rules:
    get:
      tags: [rules]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: One page of rule summaries (conditionTree omitted)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PageResponse'
                  - type: object
                    properties:
                      content:
                        type: array
                        items: { $ref: '#/components/schemas/RuleVm' }
        '400': { $ref: '#/components/responses/BadRequest' }
    post:
      tags: [rules]
      summary: Create or update a rule
      description: >
        The condition tree is compiled before persistence; an uncompilable tree is rejected rather
        than stored. A case already holding a rule rejects a second one (1:1).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RuleVm' }
      responses:
        '200': { $ref: '#/components/responses/RuleDetail' }
        '201': { $ref: '#/components/responses/RuleDetail' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413': { $ref: '#/components/responses/TooComplex' }
...
  /api/v1/rules/{ruleId}/condition:
    put:
      tags: [rule-operations]
      summary: Replace a rule's condition tree
      parameters:
        - name: ruleId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PreviewRequest' }
      responses:
        '200': { $ref: '#/components/responses/RuleDetail' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '413': { $ref: '#/components/responses/TooComplex' }
  /api/v1/rules/{ruleId}/matches:
    get:
      tags: [rule-operations]
      summary: Evaluate a stored rule
      description: >
        Returns person SUMMARY forms — nationalId is structurally absent from match results.
        Soft-deleted persons are excluded automatically by the entity-level soft-delete filter.
      parameters:
        - name: ruleId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: scope
          in: query
          schema: { type: string, enum: [GLOBAL, CASE_SCOPED], default: GLOBAL }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200': { $ref: '#/components/responses/PersonSummaryPage' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/rules/preview:
    post:
      tags: [rule-operations]
      summary: Dry-run an unsaved condition tree
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PreviewRequest' }
      responses:
        '200': { $ref: '#/components/responses/PersonSummaryPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '413': { $ref: '#/components/responses/TooComplex' }
...
components:
  parameters:
...
    Page:
      name: page
      in: query
      description: Zero-based. Negative values clamp to 0.
      schema: { type: integer, default: 0, minimum: 0 }
    Size:
      name: size
      in: query
      description: >
        Clamped server-side to [1, rule-engine.max-page-size]. An oversized request is silently
        clamped, never rejected. Omitted means rule-engine.default-page-size.
      schema: { type: integer, default: 50, minimum: 1, maximum: 500 }
    Sort:
      name: sort
      in: query
      description: >
        "field" or "field,desc". Only fields the resource declares are accepted; anything else is
        400 INVALID_SORT_FIELD. The record identifier is always appended as a tiebreaker so paging
        is totally ordered.
      schema: { type: string }
      example: name,desc
  responses:
...
    RuleDetail:
      description: Rule, detail form (includes the condition tree)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/RuleVm' }
...
    PersonSummaryPage:
      description: >
        One page of person SUMMARY responses. nationalId is not declared on that type, so no list
        path and no match path can emit it.
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/PageResponse'
              - type: object
                properties:
                  content:
                    type: array
                    items: { $ref: '#/components/schemas/PersonVm' }
    BadRequest:
      description: VALIDATION_FAILED | INVALID_SORT_FIELD | MALFORMED_REQUEST | VERSION_REQUIRED | INVALID_ARGUMENT | UNKNOWN_FIELD | INCOMPATIBLE_OPERATOR | INVALID_RULE
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: RECORD_NOT_FOUND | RULE_NOT_FOUND (a soft-deleted person is not found)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    Conflict:
      description: CONCURRENT_MODIFICATION (stale version) | CONSTRAINT_VIOLATION
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    TooComplex:
      description: RULE_TREE_TOO_COMPLEX
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
    ResourceVm:
      type: object
      description: >
        Base envelope for every UI object. One record per entity, travelling in both directions.
        Depth is a Jackson view, not a separate type: fields marked "DETAIL ONLY" below carry
        @JsonView(Vms.Detail) and are serialised only on single-record responses. Fields marked
        readOnly carry @JsonProperty(access = READ_ONLY) - emitted, but ignored on deserialisation,
        so a client cannot forge them (FR-023).
      properties:
        id:
          type: [string, 'null']
          description: Absent or null creates; present updates. The FR-003 discriminator.
        version:
          type: integer
          format: int32
          default: 0
          description: >
            Optimistic-lock token from a prior read. Omitting it means 0 - correct on create, and
            safely stale on update. A value that does not match the stored one gives 409.
        type:
          type: string
          readOnly: true
          description: Discriminator naming the record type (FR-014).
      required: [version]
    PersonVm:
      allOf:
        - $ref: '#/components/schemas/ResourceVm'
        - type: object
          properties:
            id: { type: [string, 'null'], format: uuid }
            type: { const: person, readOnly: true }
            name: { type: string, maxLength: 200 }
            age: { type: integer, minimum: 0, maximum: 149 }
            city: { type: [string, 'null'] }
            risk: { type: string, enum: [LOW, MEDIUM, HIGH, CRITICAL] }
            nationalId:
              type: string
              description: >
                DETAIL ONLY. @JsonView(Vms.Detail) keeps it off every list response and every
                rule-match result - screening should not widen exposure of a direct identifier.
            caseLinks:
              type: array
              readOnly: true
              description: DETAIL ONLY, first page.
              items: { $ref: '#/components/schemas/PersonCaseVm' }
            caseLinkCount:
              type: integer
              format: int64
              readOnly: true
              description: DETAIL ONLY. Total links, so the client knows what the page is a page of.
            createdAt: { type: string, format: date-time, readOnly: true, description: DETAIL ONLY. }
            createdBy: { type: string, readOnly: true, description: DETAIL ONLY. }
            updatedAt: { type: string, format: date-time, readOnly: true, description: DETAIL ONLY. }
            updatedBy: { type: string, readOnly: true, description: DETAIL ONLY. }
          required: [name, age, risk, nationalId]
    CaseFileVm:
      allOf:
        - $ref: '#/components/schemas/ResourceVm'
        - type: object
          properties:
            id: { type: [string, 'null'], format: uuid }
            type: { const: case, readOnly: true }
            title: { type: string, maxLength: 200 }
            status:
              type: string
              enum: [OPEN, UNDER_REVIEW, CLOSED]
              description: CLOSED is how a case is retired; there is no delete.
            openedAt: { type: string, format: date-time, readOnly: true }
            ruleId: { type: [string, 'null'], format: uuid, readOnly: true, description: DETAIL ONLY. }
            linkedPersons:
              type: array
              readOnly: true
              description: >
                DETAIL ONLY, and only the FIRST PAGE. A case accumulates an unbounded number of
                links, so the relation is never fully materialised in a response.
              items: { $ref: '#/components/schemas/PersonVm' }
            linkedPersonCount:
              type: integer
              format: int64
              readOnly: true
              description: DETAIL ONLY. Total links, so the client knows what the page is a page of.
            createdBy: { type: string, readOnly: true, description: DETAIL ONLY. }
            updatedAt: { type: string, format: date-time, readOnly: true, description: DETAIL ONLY. }
            updatedBy: { type: string, readOnly: true, description: DETAIL ONLY. }
          required: [title, status]
    RuleVm:
      allOf:
        - $ref: '#/components/schemas/ResourceVm'
        - type: object
          properties:
            id: { type: [string, 'null'], format: uuid }
            type: { const: rule, readOnly: true }
            caseId: { type: string, format: uuid, description: 'Required on create. rule:case is 1:1.' }
            name: { type: string, maxLength: 200 }
            enabled:
              type: boolean
              default: true
              description: false is how a rule is retired; there is no delete.
            condition:
              description: DETAIL ONLY - a condition AST is large and useless in a list.
              $ref: '#/components/schemas/RuleNode'
            createdAt: { type: string, format: date-time, readOnly: true, description: DETAIL ONLY. }
            createdBy: { type: string, readOnly: true, description: DETAIL ONLY. }
            updatedAt: { type: string, format: date-time, readOnly: true }
            updatedBy: { type: string, readOnly: true, description: DETAIL ONLY. }
          required: [caseId, name, condition]
    PersonCaseVm:
      allOf:
        - $ref: '#/components/schemas/ResourceVm'
        - type: object
          properties:
            id:
              type: [string, 'null']
              description: 'Composite: "<personUuid>:<caseUuid>".'
            type: { const: person-case, readOnly: true }
            personId: { type: string, format: uuid }
            caseId: { type: string, format: uuid }
            role: { type: string, enum: [SUBJECT, ASSOCIATE, WITNESS] }
            linkedAt: { type: string, format: date-time, readOnly: true }
            person:
              readOnly: true
              description: DETAIL ONLY, summary form.
              $ref: '#/components/schemas/PersonVm'
            caseFile:
              readOnly: true
              description: DETAIL ONLY, summary form.
              $ref: '#/components/schemas/CaseFileVm'
            createdBy: { type: string, readOnly: true, description: DETAIL ONLY. }
            updatedAt: { type: string, format: date-time, readOnly: true, description: DETAIL ONLY. }
            updatedBy: { type: string, readOnly: true, description: DETAIL ONLY. }
          required: [personId, caseId, role]
...
    PageResponse:
      type: object
      properties:
        content: { type: array, items: {} }
        page: { type: integer }
        size:
          type: integer
          description: The EFFECTIVE size after clamping, not what was requested.
        totalElements: { type: integer, format: int64 }
        totalPages: { type: integer }
      required: [content, page, size, totalElements, totalPages]
    PreviewRequest:
      type: object
      description: Operation payload, not a record representation - deliberately not a Vm.
      properties:
        condition: { $ref: '#/components/schemas/RuleNode' }
      required: [condition]
    RuleNode:
      type: object
      description: >
        Polymorphic condition AST - GroupNode | ConditionNode | UnaryConditionNode, discriminated by
        a "type" property. Defined by the existing rule.model classes; unchanged by this feature.
      additionalProperties: true
    ErrorResponse:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
        timestamp: { type: string, format: date-time }
      required: [code, message, timestamp]